refactor: split userdefault setting responsibilities
