+++ scripts/auth_flow_session_snapshot_unit_check.swift
@@ -21,6 +21,14 @@ func loadMany(_ relativePaths: [String]) -> String {
 
 let source = loadMany([
     "dogArea/Source/UserdefaultSetting.swift",
+    "dogArea/Source/UserDefaultsSupport/UserSessionModels.swift",
+    "dogArea/Source/UserDefaultsSupport/UserDefaultsCodableExtensions.swift",
+    "dogArea/Source/UserDefaultsSupport/UserdefaultSetting+SessionFacade.swift",
+    "dogArea/Source/UserDefaultsSupport/UserSessionStore.swift",
+    "dogArea/Source/UserDefaultsSupport/AppPreferenceStores.swift",
+    "dogArea/Source/UserDefaultsSupport/FeatureFlagStore.swift",
+    "dogArea/Source/UserDefaultsSupport/AppMetricTracker.swift",
+    "dogArea/Source/UserDefaultsSupport/SyncOutboxStore.swift",
     "dogArea/Source/AppSession/AppFeatureGate.swift",
     "dogArea/Source/AppSession/GuestDataUpgradeService.swift",
     "dogArea/Source/AppSession/AuthFlowCoordinator.swift"
