+++ scripts/ios_pr_check.sh
@@ -26,6 +26,7 @@ EXISTING_SCHEMA="
 echo "[dogArea] running document/unit checks"
 swift scripts/swift_stability_unit_check.swift
 swift scripts/userdefault_store_split_unit_check.swift
+swift scripts/userdefault_setting_second_split_unit_check.swift
 swift scripts/map_motion_pack_unit_check.swift
 swift scripts/quest_motion_pack_unit_check.swift
 swift scripts/quest_stage1_policy_unit_check.swift
